Add BigDecimal in RatesResponse

+++ src/main/java/com/example/mygifservice/models/RatesResponse.java
@@ -2,6 +2,7 @@ package com.example.mygifservice.models;
 
 import lombok.*;
 
+import java.math.BigDecimal;
 import java.sql.Timestamp;
 import java.util.Map;
 
@@ -16,6 +17,6 @@ public class RatesResponse {
     String license;
     Long timestamp;
     String base;
-    Map<String, Double> rates;
+    Map<String, BigDecimal> rates;
 
 }
